@@ -1,28 +1,29 @@
 module github.com/GoogleCloudPlatform/osconfig
 
-go 1.13
+go 1.16
 
 require (
-	cloud.google.com/go v0.82.0
-	cloud.google.com/go/logging v1.1.2 // indirect
-	cloud.google.com/go/storage v1.12.0
+	cloud.google.com/go v0.95.0
+	cloud.google.com/go/logging v1.4.2 // indirect
+	cloud.google.com/go/osconfig v0.2.0
+	cloud.google.com/go/storage v1.16.1
 	cos.googlesource.com/cos/tools.git v0.0.0-20210329212435-a349a79f950d
-	github.com/GoogleCloudPlatform/guest-logging-go v0.0.0-20210506184106-4c72f3089987
-	github.com/StackExchange/wmi v0.0.0-20190523213315-cbe66965904d
-	github.com/go-ole/go-ole v1.2.4
-	github.com/golang/mock v1.5.0
-	github.com/google/go-cmp v0.5.5
+	github.com/GoogleCloudPlatform/guest-logging-go v0.0.0-20210621170517-facc9c0d81b9
+	github.com/StackExchange/wmi v1.2.1
+	github.com/go-ole/go-ole v1.2.5
+	github.com/golang/mock v1.6.0
+	github.com/google/go-cmp v0.5.6
 	github.com/op/go-logging v0.0.0-20160315200505-970db520ece7 // indirect
-	github.com/sirupsen/logrus v1.7.0 // indirect
+	github.com/sirupsen/logrus v1.8.1 // indirect
 	github.com/tarm/serial v0.0.0-20180830185346-98f6abe2eb07
-	github.com/ulikunitz/xz v0.5.8
+	github.com/ulikunitz/xz v0.5.10
 	go.chromium.org/luci v0.0.0-20201204084249-3e81ee3e83fe // indirect
-	golang.org/x/crypto v0.0.0-20201203163018-be400aefbc4c
-	golang.org/x/net v0.0.0-20210521195947-fe42d452be8f // indirect
-	golang.org/x/oauth2 v0.0.0-20210514164344-f6687ab2804c
-	golang.org/x/sys v0.0.0-20210521203332-0cec03c779c1
-	google.golang.org/api v0.46.0
-	google.golang.org/genproto v0.0.0-20210517163617-5e0236093d7a
-	google.golang.org/grpc v1.37.1
-	google.golang.org/protobuf v1.26.0
+	golang.org/x/crypto v0.0.0-20210921155107-089bfa567519
+	golang.org/x/net v0.0.0-20210917221730-978cfadd31cf // indirect
+	golang.org/x/oauth2 v0.0.0-20210819190943-2bc19b11175f
+	golang.org/x/sys v0.0.0-20210923061019-b8560ed6a9b7
+	google.golang.org/api v0.57.0
+	google.golang.org/genproto v0.0.0-20210923133937-385e58f4e555
+	google.golang.org/grpc v1.40.0
+	google.golang.org/protobuf v1.27.1
 )
