@@ -12,7 +12,8 @@
 //  See the License for the specific language governing permissions and
 //  limitations under the License.
 
-//+build !test
+//go:build !test
+// +build !test
 
 package ospatch
 
@@ -188,7 +189,7 @@ func GetWUAUpdates(ctx context.Context, session *packages.IUpdateSession, classF
 	if err != nil {
 		return nil, fmt.Errorf("GetWUAUpdateCollection error: %v", err)
 	}
-	if len(classFilter) == 0 && len(kbExcludes) == 0 {
+	if len(classFilter) == 0 && len(kbExcludes) == 0 && len(exclusivePatches) == 0 {
 		return updts, nil
 	}
 	defer updts.Release()
