@@ -30,6 +30,7 @@ import (
 	agentendpoint "cloud.google.com/go/osconfig/agentendpoint/apiv1"
 	"github.com/GoogleCloudPlatform/osconfig/agentconfig"
 	"github.com/GoogleCloudPlatform/osconfig/clog"
+	"github.com/GoogleCloudPlatform/osconfig/osinfo"
 	"github.com/GoogleCloudPlatform/osconfig/retryutil"
 	"github.com/GoogleCloudPlatform/osconfig/tasker"
 	"google.golang.org/api/option"
@@ -49,6 +50,7 @@ var (
 	errServiceNotEnabled = errors.New("service is not enabled for this project")
 	errResourceExhausted = errors.New("ResourceExhausted")
 	taskStateFile        = agentconfig.TaskStateFile()
+	oldTaskStateFile     = agentconfig.OldTaskStateFile()
 	sameStateTimeWindow  = -5 * time.Second
 )
 
@@ -101,12 +103,32 @@ func (c *Client) RegisterAgent(ctx context.Context) error {
 		return err
 	}
 
-	req := &agentendpointpb.RegisterAgentRequest{AgentVersion: agentconfig.Version(), SupportedCapabilities: agentconfig.Capabilities()}
+	oi := &osinfo.OSInfo{}
+	if agentconfig.OSInventoryEnabled() {
+		oi, err = osinfo.Get()
+		if err != nil {
+			// Log the error but still call RegisterAgent (fields will be empty).
+			clog.Errorf(ctx, "osinfo.Get() error: %v", err)
+		}
+	}
+
+	req := &agentendpointpb.RegisterAgentRequest{
+		AgentVersion:          agentconfig.Version(),
+		SupportedCapabilities: agentconfig.Capabilities(),
+		OsLongName:            oi.LongName,
+		OsShortName:           oi.ShortName,
+		OsVersion:             oi.Version,
+		OsArchitecture:        oi.Architecture,
+	}
 	req.InstanceIdToken = "<redacted>"
 	clog.DebugRPC(ctx, "RegisterAgent", req, nil)
 	req.InstanceIdToken = token
 
-	resp, err := c.raw.RegisterAgent(ctx, req)
+	var resp *agentendpointpb.RegisterAgentResponse
+	err = retryutil.RetryAPICall(ctx, apiRetrySec*time.Second, "RegisterAgent", func() error {
+		resp, err = c.raw.RegisterAgent(ctx, req)
+		return err
+	})
 	clog.DebugRPC(ctx, "RegisterAgent", nil, resp)
 
 	return err
@@ -221,7 +243,7 @@ func (c *Client) runTask(ctx context.Context) {
 		}
 
 		clog.Debugf(ctx, "Received task: %s.", task.GetTaskType())
-		ctx = clog.WithLabels(ctx, map[string]string{"task_type": task.GetTaskType().String()})
+		ctx := clog.WithLabels(ctx, map[string]string{"task_type": task.GetTaskType().String()})
 		switch task.GetTaskType() {
 		case agentendpointpb.TaskType_APPLY_PATCHES:
 			if err := c.RunApplyPatches(ctx, task); err != nil {
