@@ -19,6 +19,7 @@ import (
 	"io/ioutil"
 	"os"
 	"path/filepath"
+	"runtime"
 )
 
 type taskState struct {
@@ -47,14 +48,25 @@ func (s *taskState) save(path string) error {
 }
 
 func loadState(path string) (*taskState, error) {
+	// We load the current state file first, if it does not exist we try to load the old state file.
 	d, err := ioutil.ReadFile(path)
 	if os.IsNotExist(err) {
-		return nil, nil
+		if runtime.GOOS == "windows" {
+			return nil, nil
+		}
+		d, err = ioutil.ReadFile(oldTaskStateFile)
+		if os.IsNotExist(err) {
+			return nil, nil
+		}
 	}
 	if err != nil {
 		return nil, err
 	}
 
+	// Cleanup old state file if needed.
+	if runtime.GOOS != "windows" {
+		os.Remove(oldTaskStateFile)
+	}
 	var st taskState
 	return &st, json.Unmarshal(d, &st)
 }
