@@ -16,6 +16,8 @@ package config
 
 import (
 	"context"
+	"encoding/base64"
+	"encoding/json"
 	"fmt"
 	"io/ioutil"
 	"os"
@@ -23,11 +25,20 @@ import (
 	"strings"
 	"time"
 
+	"github.com/GoogleCloudPlatform/osconfig/agentconfig"
 	"github.com/GoogleCloudPlatform/osconfig/clog"
 	"github.com/GoogleCloudPlatform/osconfig/packages"
 	"github.com/GoogleCloudPlatform/osconfig/util"
 
 	agentendpointpb "google.golang.org/genproto/googleapis/cloud/osconfig/agentendpoint/v1"
+	"google.golang.org/protobuf/proto"
+)
+
+var (
+	packageInfoCacheFile = filepath.Join(agentconfig.CacheDir(), "config_package_info.cache")
+	// Clear out the entry if the last lookup is > 7 days ago.
+	packageInfoCacheTimeout = -168 * time.Hour
+	packageInfoCacheStore   packageInfoCache
 )
 
 type packageResouce struct {
@@ -56,8 +67,8 @@ type GooGetPackage struct {
 
 // MSIPackage describes an msi package resource.
 type MSIPackage struct {
-	PackageResource        *agentendpointpb.OSPolicy_Resource_PackageResource_MSI
-	productName, localPath string
+	PackageResource                     *agentendpointpb.OSPolicy_Resource_PackageResource_MSI
+	productName, productCode, localPath string
 }
 
 // YumPackage describes a yum package resource.
@@ -100,6 +111,88 @@ func (p *packageResouce) validateFile(file *agentendpointpb.OSPolicy_Resource_Fi
 	return nil
 }
 
+type packageInfoCache map[string]packageInfo
+
+type packageInfo struct {
+	PkgInfo    *packages.PkgInfo
+	LastLookup time.Time
+}
+
+func getPackageInfoCacheKey(pkgFile *agentendpointpb.OSPolicy_Resource_File) (string, error) {
+	// We use the base64 encoded binary proto as the key.
+	raw, err := proto.Marshal(pkgFile)
+	if err != nil {
+		return "", err
+	}
+	return base64.RawStdEncoding.EncodeToString(raw), nil
+}
+
+func loadPackageInfoCache(ctx context.Context) {
+	if packageInfoCacheStore != nil {
+		return
+	}
+	data, err := ioutil.ReadFile(packageInfoCacheFile)
+	if err != nil {
+		// Just ignore the error and return early
+		// The error mode here is to just always redownload the file.
+		clog.Debugf(ctx, "Error reading the package info cache: %v", err)
+		packageInfoCacheStore = packageInfoCache{}
+		return
+	}
+	var cache packageInfoCache
+	if err := json.Unmarshal(data, &cache); err != nil {
+		clog.Debugf(ctx, "Error unmarshaling the package info cache: %v", err)
+		packageInfoCacheStore = packageInfoCache{}
+		return
+	}
+	packageInfoCacheStore = cache
+}
+
+func getPackageInfoFromCache(ctx context.Context, pkgFile *agentendpointpb.OSPolicy_Resource_File) *packages.PkgInfo {
+	loadPackageInfoCache(ctx)
+	key, err := getPackageInfoCacheKey(pkgFile)
+	if err != nil {
+		// Just ignore the error and return early
+		// The error mode here is just always redownload the file.
+		clog.Debugf(ctx, "Error creating the package info cache key: %v", err)
+		return nil
+	}
+	packageInfo, ok := packageInfoCacheStore[key]
+	if !ok {
+		return nil
+	}
+	return packageInfo.PkgInfo
+}
+
+func updatePackageInfoCache(ctx context.Context, info *packages.PkgInfo, pkgFile *agentendpointpb.OSPolicy_Resource_File) {
+	loadPackageInfoCache(ctx)
+	for k, v := range packageInfoCacheStore {
+		if time.Now().Add(packageInfoCacheTimeout).After(v.LastLookup) {
+			delete(packageInfoCacheStore, k)
+		}
+	}
+	key, err := getPackageInfoCacheKey(pkgFile)
+	if err != nil {
+		// Just ignore the error and return early
+		// The error mode here is to just always redownload the file.
+		clog.Warningf(ctx, "Error creating the package info cache: %v", err)
+		return
+	}
+	packageInfoCacheStore[key] = packageInfo{PkgInfo: info, LastLookup: time.Now()}
+}
+
+func savePackageInfoCache(ctx context.Context) error {
+	if packageInfoCacheStore == nil {
+		return nil
+	}
+	data, err := json.Marshal(packageInfoCacheStore)
+	if err != nil {
+		return err
+	}
+	packageInfoCacheStore = nil
+	return ioutil.WriteFile(packageInfoCacheFile, data, 0644)
+}
+
 func (p *packageResouce) validate(ctx context.Context) (*ManagedResources, error) {
 	switch p.GetSystemPackage().(type) {
 	case *agentendpointpb.OSPolicy_Resource_PackageResource_Apt:
@@ -121,14 +214,22 @@ func (p *packageResouce) validate(ctx context.Context) (*ManagedResources, error
 		if err := p.validateFile(pr.GetSource()); err != nil {
 			return nil, err
 		}
-		localPath, err := p.download(ctx, "pkg.deb", p.GetDeb().GetSource())
-		if err != nil {
-			return nil, err
+		var localPath string
+		var err error
+		source := p.GetDeb().GetSource()
+		info := getPackageInfoFromCache(ctx, source)
+		if info == nil {
+			localPath, err = p.download(ctx, "pkg.deb", source)
+			if err != nil {
+				return nil, err
+			}
+			info, err = packages.DebPkgInfo(ctx, localPath)
+			if err != nil {
+				return nil, err
+			}
 		}
-		info, err := packages.DebPkgInfo(ctx, localPath)
-		if err != nil {
-			return nil, err
-		}
+		// Always update the cache to update the timestamps.
+		updatePackageInfoCache(ctx, info, source)
 
 		p.managedPackage.Deb = &DebPackage{PackageResource: pr, localPath: localPath, name: info.Name}
 
@@ -151,12 +252,26 @@ func (p *packageResouce) validate(ctx context.Context) (*ManagedResources, error
 		if err := p.validateFile(pr.GetSource()); err != nil {
 			return nil, err
 		}
-		localPath, err := p.download(ctx, "pkg.msi", p.GetMsi().GetSource())
-		if err != nil {
-			return nil, err
+		var localPath string
+		var err error
+		source := p.GetMsi().GetSource()
+		info := getPackageInfoFromCache(ctx, source)
+		if info == nil {
+			localPath, err = p.download(ctx, "pkg.msi", source)
+			if err != nil {
+				return nil, err
+			}
+			productName, productCode, err := packages.MSIInfo(localPath)
+			if err != nil {
+				return nil, err
+			}
+			// We store productCode as version in the packageinfo struct.
+			info = &packages.PkgInfo{Name: productName, Version: productCode}
 		}
+		// Always update the cache to update the timestamps.
+		updatePackageInfoCache(ctx, info, source)
 
-		p.managedPackage.MSI = &MSIPackage{PackageResource: pr, localPath: localPath}
+		p.managedPackage.MSI = &MSIPackage{PackageResource: pr, localPath: localPath, productName: info.Name, productCode: info.Version}
 
 	case *agentendpointpb.OSPolicy_Resource_PackageResource_Yum:
 		pr := p.GetYum()
@@ -185,14 +300,22 @@ func (p *packageResouce) validate(ctx context.Context) (*ManagedResources, error
 		if err := p.validateFile(pr.GetSource()); err != nil {
 			return nil, err
 		}
-		localPath, err := p.download(ctx, "pkg.rpm", p.GetRpm().GetSource())
-		if err != nil {
-			return nil, err
+		var localPath string
+		var err error
+		source := p.GetRpm().GetSource()
+		info := getPackageInfoFromCache(ctx, source)
+		if info == nil {
+			localPath, err = p.download(ctx, "pkg.rpm", source)
+			if err != nil {
+				return nil, err
+			}
+			info, err = packages.RPMPkgInfo(ctx, localPath)
+			if err != nil {
+				return nil, err
+			}
 		}
-		info, err := packages.RPMPkgInfo(ctx, localPath)
-		if err != nil {
-			return nil, err
-		}
+		// Always update the cache to update the timestamps.
+		updatePackageInfoCache(ctx, info, source)
 
 		p.managedPackage.RPM = &RPMPackage{PackageResource: pr, localPath: localPath, name: info.Name}
 
@@ -319,7 +442,7 @@ func (p *packageResouce) checkState(ctx context.Context) (inDesiredState bool, e
 
 	case p.managedPackage.MSI != nil:
 		desiredState = agentendpointpb.OSPolicy_Resource_PackageResource_INSTALLED
-		p.managedPackage.MSI.productName, pkgIns, err = packages.MSIInfo(p.managedPackage.MSI.localPath)
+		pkgIns, err = packages.MSIInstalled(p.managedPackage.MSI.productCode)
 		if err != nil {
 			return false, err
 		}
@@ -389,6 +512,14 @@ func (p *packageResouce) enforceState(ctx context.Context) (inDesiredState bool,
 		enforcePackage.packageType = "deb"
 		enforcePackage.installedCache = debInstalled
 		enforcePackage.action = installing
+		// Check if we have not pulled the package yet.
+		if p.managedPackage.Deb.localPath == "" {
+			localPath, err := p.download(ctx, "pkg.deb", p.GetDeb().GetSource())
+			if err != nil {
+				return false, err
+			}
+			p.managedPackage.Deb.localPath = localPath
+		}
 		if p.GetDeb().GetPullDeps() {
 			enforcePackage.actionFunc = func() error { return packages.InstallAptPackages(ctx, []string{p.managedPackage.Deb.localPath}) }
 		} else {
@@ -411,6 +542,14 @@ func (p *packageResouce) enforceState(ctx context.Context) (inDesiredState bool,
 		enforcePackage.packageType = "msi"
 		enforcePackage.action = installing
 		enforcePackage.installedCache = &packageCache{} // No package cache for msi.
+		// Check if we have not pulled the package yet.
+		if p.managedPackage.MSI.localPath == "" {
+			localPath, err := p.download(ctx, "pkg.msi", p.GetMsi().GetSource())
+			if err != nil {
+				return false, err
+			}
+			p.managedPackage.MSI.localPath = localPath
+		}
 		enforcePackage.actionFunc = func() error {
 			return packages.InstallMSIPackage(ctx, p.managedPackage.MSI.localPath, p.managedPackage.MSI.PackageResource.GetProperties())
 		}
@@ -442,6 +581,14 @@ func (p *packageResouce) enforceState(ctx context.Context) (inDesiredState bool,
 		enforcePackage.packageType = "rpm"
 		enforcePackage.installedCache = rpmInstalled
 		enforcePackage.action = installing
+		// Check if we have not pulled the package yet.
+		if p.managedPackage.RPM.localPath == "" {
+			localPath, err := p.download(ctx, "pkg.rpm", p.GetRpm().GetSource())
+			if err != nil {
+				return false, err
+			}
+			p.managedPackage.RPM.localPath = localPath
+		}
 		if p.GetRpm().GetPullDeps() {
 			switch {
 			case packages.YumExists:
@@ -457,7 +604,7 @@ func (p *packageResouce) enforceState(ctx context.Context) (inDesiredState bool,
 	}
 
 	clog.Infof(ctx, "%s %s package %q", strings.Title(enforcePackage.action), enforcePackage.packageType, enforcePackage.name)
-	// Reset the cache as we are taking action.
+	// Reset the cache as we are taking action on.
 	enforcePackage.installedCache.cache = nil
 	if err := enforcePackage.actionFunc(); err != nil {
 		return false, fmt.Errorf("error %s %s package %q", enforcePackage.action, enforcePackage.packageType, enforcePackage.name)
@@ -469,6 +616,10 @@ func (p *packageResouce) enforceState(ctx context.Context) (inDesiredState bool,
 func (p *packageResouce) populateOutput(rCompliance *agentendpointpb.OSPolicyResourceCompliance) {}
 
 func (p *packageResouce) cleanup(ctx context.Context) error {
+	// Save cache and clear the variable.
+	if err := savePackageInfoCache(ctx); err != nil {
+		clog.Warningf(ctx, "Error saving the package info cache: %v", err)
+	}
 	if p.managedPackage.tempDir != "" {
 		return os.RemoveAll(p.managedPackage.tempDir)
 	}
