@@ -16,10 +16,12 @@ package config
 
 import (
 	"context"
+	"encoding/json"
 	"io/ioutil"
 	"os"
 	"os/exec"
 	"path/filepath"
+	"reflect"
 	"testing"
 	"time"
 
@@ -142,7 +144,7 @@ func TestPackageResourceValidate(t *testing.T) {
 				PackageResource: &agentendpointpb.OSPolicy_Resource_PackageResource_Deb{
 					Source: &agentendpointpb.OSPolicy_Resource_File{
 						Type: &agentendpointpb.OSPolicy_Resource_File_LocalPath{LocalPath: tmpFile}}}}},
-			exec.Command("/usr/bin/dpkg-deb", "-I", tmpFile),
+			exec.CommandContext(context.Background(), "/usr/bin/dpkg-deb", "-I", tmpFile),
 			[]byte("Package: foo\nVersion: 1:1dummy-g1\nArchitecture: amd64"),
 		},
 		{
@@ -237,7 +239,7 @@ func TestPackageResourceValidate(t *testing.T) {
 				PackageResource: &agentendpointpb.OSPolicy_Resource_PackageResource_RPM{
 					Source: &agentendpointpb.OSPolicy_Resource_File{
 						Type: &agentendpointpb.OSPolicy_Resource_File_LocalPath{LocalPath: tmpFile}}}}},
-			exec.Command("/usr/bin/rpmquery", "--queryformat", "%{NAME} %{ARCH} %|EPOCH?{%{EPOCH}:}:{}|%{VERSION}-%{RELEASE}\n", "-p", tmpFile),
+			exec.CommandContext(context.Background(), "/usr/bin/rpmquery", "--queryformat", "%{NAME} %{ARCH} %|EPOCH?{%{EPOCH}:}:{}|%{VERSION}-%{RELEASE}\n", "-p", tmpFile),
 			[]byte("foo x86_64 1.2.3-4"),
 		},
 	}
@@ -248,6 +250,7 @@ func TestPackageResourceValidate(t *testing.T) {
 					ResourceType: &agentendpointpb.OSPolicy_Resource_Pkg{Pkg: tt.prpb},
 				},
 			}
+			defer pr.Cleanup(ctx)
 
 			if tt.expectedCmd != nil {
 				mockCommandRunner.EXPECT().Run(ctx, tt.expectedCmd).Return(tt.expectedCmdReturn, nil, nil)
@@ -285,7 +288,7 @@ func TestPopulateInstalledCache(t *testing.T) {
 
 	mockCommandRunner := utilmocks.NewMockCommandRunner(mockCtrl)
 	packages.SetCommandRunner(mockCommandRunner)
-	mockCommandRunner.EXPECT().Run(ctx, exec.Command("googet.exe", "installed")).Return([]byte("Installed Packages:\nfoo.x86_64 1.2.3@4\nbar.noarch 1.2.3@4"), nil, nil).Times(1)
+	mockCommandRunner.EXPECT().Run(ctx, exec.CommandContext(context.Background(), "googet.exe", "installed")).Return([]byte("Installed Packages:\nfoo.x86_64 1.2.3@4\nbar.noarch 1.2.3@4"), nil, nil).Times(1)
 
 	if err := populateInstalledCache(ctx, ManagedPackage{GooGet: &GooGetPackage{}}); err != nil {
 		t.Fatalf("Unexpected error from populateInstalledCache: %v", err)
@@ -367,6 +370,7 @@ func TestPackageResourceCheckState(t *testing.T) {
 					ResourceType: &agentendpointpb.OSPolicy_Resource_Pkg{Pkg: tt.prpb},
 				},
 			}
+			defer pr.Cleanup(ctx)
 			// Run validate first to make sure everything gets setup correctly.
 			// This adds complexity to this 'unit' test and turns it into more
 			// of a integration test but reduces overall test functions and gives
@@ -407,8 +411,8 @@ func TestPackageResourceEnforceState(t *testing.T) {
 			aptInstalledPR,
 			aptInstalled,
 			func() []*exec.Cmd {
-				cmd1 := exec.Command("/usr/bin/apt-get", "update")
-				cmd2 := exec.Command("/usr/bin/apt-get", "install", "-y", "foo")
+				cmd1 := exec.CommandContext(context.Background(), "/usr/bin/apt-get", "update")
+				cmd2 := exec.CommandContext(context.Background(), "/usr/bin/apt-get", "install", "-y", "foo")
 				cmd2.Env = append(os.Environ(),
 					"DEBIAN_FRONTEND=noninteractive",
 				)
@@ -420,7 +424,7 @@ func TestPackageResourceEnforceState(t *testing.T) {
 			aptRemovedPR,
 			aptInstalled,
 			func() []*exec.Cmd {
-				cmd1 := exec.Command("/usr/bin/apt-get", "remove", "-y", "foo")
+				cmd1 := exec.CommandContext(context.Background(), "/usr/bin/apt-get", "remove", "-y", "foo")
 				cmd1.Env = append(os.Environ(),
 					"DEBIAN_FRONTEND=noninteractive",
 				)
@@ -431,37 +435,37 @@ func TestPackageResourceEnforceState(t *testing.T) {
 			"GooGetInstalled",
 			googetInstalledPR,
 			gooInstalled,
-			[]*exec.Cmd{exec.Command("googet.exe", "-noconfirm", "install", "foo")},
+			[]*exec.Cmd{exec.CommandContext(context.Background(), "googet.exe", "-noconfirm", "install", "foo")},
 		},
 		{
 			"GooGetRemoved",
 			googetRemovedPR,
 			gooInstalled,
-			[]*exec.Cmd{exec.Command("googet.exe", "-noconfirm", "remove", "foo")},
+			[]*exec.Cmd{exec.CommandContext(context.Background(), "googet.exe", "-noconfirm", "remove", "foo")},
 		},
 		{
 			"YumInstalled",
 			yumInstalledPR,
 			yumInstalled,
-			[]*exec.Cmd{exec.Command("/usr/bin/yum", "install", "--assumeyes", "foo")},
+			[]*exec.Cmd{exec.CommandContext(context.Background(), "/usr/bin/yum", "install", "--assumeyes", "foo")},
 		},
 		{
 			"YumRemoved",
 			yumRemovedPR,
 			yumInstalled,
-			[]*exec.Cmd{exec.Command("/usr/bin/yum", "remove", "--assumeyes", "foo")},
+			[]*exec.Cmd{exec.CommandContext(context.Background(), "/usr/bin/yum", "remove", "--assumeyes", "foo")},
 		},
 		{
 			"ZypperInstalled",
 			zypperInstalledPR,
 			zypperInstalled,
-			[]*exec.Cmd{exec.Command("/usr/bin/zypper", "--gpg-auto-import-keys", "--non-interactive", "install", "--auto-agree-with-licenses", "foo")},
+			[]*exec.Cmd{exec.CommandContext(context.Background(), "/usr/bin/zypper", "--gpg-auto-import-keys", "--non-interactive", "install", "--auto-agree-with-licenses", "foo")},
 		},
 		{
 			"ZypperRemoved",
 			zypperRemovedPR,
 			zypperInstalled,
-			[]*exec.Cmd{exec.Command("/usr/bin/zypper", "--non-interactive", "remove", "foo")},
+			[]*exec.Cmd{exec.CommandContext(context.Background(), "/usr/bin/zypper", "--non-interactive", "remove", "foo")},
 		},
 	}
 
@@ -472,6 +476,8 @@ func TestPackageResourceEnforceState(t *testing.T) {
 					ResourceType: &agentendpointpb.OSPolicy_Resource_Pkg{Pkg: tt.prpb},
 				},
 			}
+			defer pr.Cleanup(ctx)
+
 			// Run Validate first to make sure everything gets setup correctly.
 			// This adds complexity to this 'unit' test and turns it into more
 			// of a integration test but reduces overall test functions and gives
@@ -494,5 +500,72 @@ func TestPackageResourceEnforceState(t *testing.T) {
 				t.Errorf("Enforce function did not set package cache to nil")
 			}
 		})
+	}
+}
+
+func TestPackageInfoCache(t *testing.T) {
+	ctx := context.Background()
+	pkgInfo := &packages.PkgInfo{Name: "name", Arch: "arch", Version: "version"}
+	pkgFile := &agentendpointpb.OSPolicy_Resource_File{AllowInsecure: true, Type: &agentendpointpb.OSPolicy_Resource_File_Gcs_{Gcs: &agentendpointpb.OSPolicy_Resource_File_Gcs{Bucket: "bucket", Object: "object", Generation: 123456789}}}
+	wantKey := "IAESFQoGYnVja2V0EgZvYmplY3QYlZrvOg"
+
+	tmpDir, err := ioutil.TempDir("", "")
+	if err != nil {
+		t.Fatal(err)
+	}
+	defer os.RemoveAll(tmpDir)
+	packageInfoCacheFile = filepath.Join(tmpDir, "file.cache")
+	packageInfoCacheStore = nil
+
+	updatePackageInfoCache(ctx, pkgInfo, pkgFile)
+	if err := savePackageInfoCache(ctx); err != nil {
+		t.Fatal(err)
+	}
+	got := getPackageInfoFromCache(ctx, pkgFile)
+	if !reflect.DeepEqual(got, pkgInfo) {
+		t.Errorf("Did not get expected cache data, got: %+v, want: %+v", got, pkgInfo)
+	}
+	if _, ok := packageInfoCacheStore[wantKey]; !ok {
+		t.Errorf("Cache did not contain expected key, cache: %+v, want: %q", packageInfoCacheStore, wantKey)
+	}
+
+	// Now test save and reload.
+	savePackageInfoCache(ctx)
+	if packageInfoCacheStore != nil {
+		t.Fatal("expected packageInfoCacheStore to be nil")
+	}
+	loadPackageInfoCache(ctx)
+	if _, ok := packageInfoCacheStore[wantKey]; !ok {
+		t.Errorf("Cache did not contain expected key, cache: %+v, want: %q", packageInfoCacheStore, wantKey)
+	}
+}
+
+func TestUpdatePackageInfoCacheTimeout(t *testing.T) {
+	ctx := context.Background()
+
+	tmpDir, err := ioutil.TempDir("", "")
+	if err != nil {
+		t.Fatal(err)
+	}
+	defer os.RemoveAll(tmpDir)
+	packageInfoCacheFile = filepath.Join(tmpDir, "file.cache")
+	packageInfoCacheStore = nil
+
+	cache := packageInfoCache{}
+	key := "IAESFQoGYnVja2V0EgZvYmplY3QYlZrvOg"
+	info := &packages.PkgInfo{Name: "name", Arch: "arch", Version: "version"}
+	cache[key] = packageInfo{PkgInfo: info, LastLookup: time.Now().Add(packageInfoCacheTimeout).Add(-1 * time.Hour)}
+
+	data, err := json.Marshal(cache)
+	if err != nil {
+		t.Fatal(err)
+	}
+	if err := ioutil.WriteFile(packageInfoCacheFile, data, 0644); err != nil {
+		t.Fatal(err)
+	}
+
+	updatePackageInfoCache(ctx, nil, nil)
+	if _, ok := packageInfoCacheStore[key]; ok {
+		t.Errorf("Cache should not contain expired data, cache: %+v", packageInfoCacheStore)
 	}
 }
