@@ -20,6 +20,7 @@ import (
 	"flag"
 	"fmt"
 	"io"
+	"io/ioutil"
 	"math/rand"
 	"net/http"
 	"os"
@@ -35,6 +36,7 @@ import (
 	"github.com/GoogleCloudPlatform/osconfig/clog"
 	"github.com/GoogleCloudPlatform/osconfig/policies"
 	"github.com/GoogleCloudPlatform/osconfig/tasker"
+	"github.com/GoogleCloudPlatform/osconfig/util"
 	"github.com/tarm/serial"
 
 	_ "net/http/pprof"
@@ -74,19 +76,29 @@ func (s *serialPort) Write(b []byte) (int, error) {
 
 var deferredFuncs []func()
 
+// RegisterAgent is a blocking call, the RPC itself has retry logic baked in
+// with jitter and backoff up to a total of 10 minutes.
+// If client creation or register agent (after retries) fail we then wait for
+// 5 minutes and try again.
 func registerAgent(ctx context.Context) {
-	if agentconfig.TaskNotificationEnabled() || agentconfig.GuestPoliciesEnabled() {
+	for {
 		if client, err := agentendpoint.NewClient(ctx); err != nil {
 			logger.Errorf(err.Error())
 		} else if err := client.RegisterAgent(ctx); err != nil {
 			logger.Errorf(err.Error())
+			client.Close()
+		} else {
+			// RegisterAgent completed successfully.
+			client.Close()
+			return
 		}
+		time.Sleep(5 * time.Minute)
 	}
 }
 
 func run(ctx context.Context) {
 	// Setup logging.
-	opts := logger.LogOpts{LoggerName: "OSConfigAgent", UserAgent: agentconfig.UserAgent()}
+	opts := logger.LogOpts{LoggerName: "OSConfigAgent", UserAgent: agentconfig.UserAgent(), DisableLocalLogging: agentconfig.DisableLocalLogging()}
 	if agentconfig.Stdout() {
 		opts.Writers = []io.Writer{os.Stdout}
 	}
@@ -114,6 +126,21 @@ func run(ctx context.Context) {
 		clog.Errorf(ctx, "Error removing restart signal file: %v", err)
 	}
 
+	// On shutdown if the old restart file exists, and there is nothing else in that old directory,
+	// cleanup that directory ignoring all errors.
+	// This ensures we only cleanup this directory if we were using it with an old version of the agent.
+	deferredFuncs = append(deferredFuncs, func() {
+		if runtime.GOOS == "linux" && util.Exists(agentconfig.OldRestartFile()) {
+			os.Remove(agentconfig.OldRestartFile())
+
+			files, err := ioutil.ReadDir(filepath.Dir(agentconfig.OldRestartFile()))
+			if err != nil || len(files) > 0 {
+				return
+			}
+			os.RemoveAll(filepath.Dir(agentconfig.OldRestartFile()))
+		}
+	})
+
 	deferredFuncs = append(deferredFuncs, logger.Close, func() { clog.Infof(ctx, "OSConfig Agent (version %s) shutting down.", agentconfig.Version()) })
 
 	obtainLock()
@@ -123,12 +150,14 @@ func run(ctx context.Context) {
 
 	clog.Infof(ctx, "OSConfig Agent (version %s) started.", agentconfig.Version())
 
-	// Call RegisterAgent on start then at least once every day.
-	registerAgent(ctx)
+	// Call RegisterAgent at least once every day, on start calling
+	// of RegisterAgent is handled in the service loop.
 	go func() {
 		c := time.Tick(24 * time.Hour)
 		for range c {
-			registerAgent(ctx)
+			if agentconfig.TaskNotificationEnabled() || agentconfig.GuestPoliciesEnabled() {
+				registerAgent(ctx)
+			}
 		}
 	}()
 
@@ -170,6 +199,11 @@ func runTaskLoop(ctx context.Context, c chan struct{}) {
 		// Set debug logging settings so that customers don't need to restart the agent.
 		logger.SetDebugLogging(agentconfig.Debug())
 		clog.DebugEnabled = agentconfig.Debug()
+		if agentconfig.TaskNotificationEnabled() && taskNotificationClient == nil {
+			// Call RegisterAgent now since we just either started running or were just enabled.
+			// This call is blocking until successful as we can't continue unless register agent has completed.
+			registerAgent(ctx)
+		}
 		if agentconfig.TaskNotificationEnabled() && (taskNotificationClient == nil || taskNotificationClient.Closed()) {
 			// Start WaitForTaskNotification if we need to.
 			taskNotificationClient, err = agentendpoint.NewClient(ctx)
@@ -192,6 +226,7 @@ func runTaskLoop(ctx context.Context, c chan struct{}) {
 		default:
 		}
 
+		// Wait on any metadata config change.
 		if err := agentconfig.WatchConfig(ctx); err != nil {
 			clog.Errorf(ctx, err.Error())
 		}
@@ -204,19 +239,10 @@ func runTaskLoop(ctx context.Context, c chan struct{}) {
 	}
 }
 
-func runServiceLoop(ctx context.Context) {
-	// This is just to ensure WaitForTaskNotification runs before any periodocs.
-	c := make(chan struct{})
-	// Configures WaitForTaskNotification, waits for config changes with WatchConfig.
-	go runTaskLoop(ctx, c)
-	<-c
-
-	// Runs functions that need to run on a set interval.
-	ticker := time.NewTicker(agentconfig.SvcPollInterval())
+// Runs internal functions that need to run on an interval.
+func runInternalPeriodics(ctx context.Context) {
+	ticker := time.NewTicker(10 * time.Minute)
 	defer ticker.Stop()
-	// First inventory run will be somewhere between 3 and 5 min.
-	firstInventory := time.After(time.Duration(rand.Intn(120)+180) * time.Second)
-	ranFirstInventory := false
 	for {
 		if _, err := os.Stat(agentconfig.RestartFile()); err == nil {
 			clog.Infof(ctx, "Restart required marker file exists, beginning agent shutdown, waiting for tasks to complete.")
@@ -228,6 +254,32 @@ func runServiceLoop(ctx context.Context) {
 			os.Exit(2)
 		}
 
+		select {
+		case <-ticker.C:
+			continue
+		case <-ctx.Done():
+			return
+		}
+	}
+}
+
+func runServiceLoop(ctx context.Context) {
+	go runInternalPeriodics(ctx)
+
+	// This is just to ensure WaitForTaskNotification runs before any other tasks.
+	c := make(chan struct{})
+	// Configures WaitForTaskNotification, waits for config changes with WatchConfig.
+	go runTaskLoop(ctx, c)
+	// Don't continue any other tasks until WaitForTaskNotification has run.
+	<-c
+
+	// Runs functions that need to run on a set interval.
+	ticker := time.NewTicker(agentconfig.SvcPollInterval())
+	defer ticker.Stop()
+	// First inventory run will be somewhere between 3 and 5 min.
+	firstInventory := time.After(time.Duration(rand.Intn(120)+180) * time.Second)
+	ranFirstInventory := false
+	for {
 		if agentconfig.GuestPoliciesEnabled() {
 			policies.Run(ctx)
 		}
