@@ -116,6 +116,31 @@ func dpkgRepair(ctx context.Context, out []byte) bool {
 	return true
 }
 
+type cmdModifier func(*exec.Cmd)
+
+func runAptGet(ctx context.Context, args []string, cmdModifiers []cmdModifier) ([]byte, []byte, error) {
+	cmd := exec.CommandContext(ctx, aptGet, args...)
+	for _, modifier := range cmdModifiers {
+		modifier(cmd)
+	}
+
+	return runner.Run(ctx, cmd)
+}
+
+func runAptGetWithDowngradeRetrial(ctx context.Context, args []string, cmdModifiers []cmdModifier) ([]byte, []byte, error) {
+	stdout, stderr, err := runAptGet(ctx, args, cmdModifiers)
+	if err != nil {
+		if strings.Contains(string(stderr), "E: Packages were downgraded and -y was used without --allow-downgrades.") {
+			cmdModifiers = append(cmdModifiers, func(cmd *exec.Cmd) {
+				cmd.Args = append(cmd.Args, allowDowngradesArg)
+			})
+			stdout, stderr, err = runAptGet(ctx, args, cmdModifiers)
+		}
+	}
+
+	return stdout, stderr, err
+}
+
 func parseDpkgDeb(data []byte) (*PkgInfo, error) {
 	/*
 	   new Debian package, version 2.0.
@@ -150,7 +175,11 @@ func parseDpkgDeb(data []byte) (*PkgInfo, error) {
 			continue
 		}
 		if bytes.Contains(fields[0], []byte("Package:")) {
-			info.Name = string(fields[1])
+			// Some packages do not adhere to the Debian Policy and might have mix-cased names
+			// And dpkg will register the package with lower case anyway so use lower-case package name
+			// This is necessary because the compliance check is done between the .deb file descriptor value
+			// and the internal dpkg db which register a lower-cased package name
+			info.Name = strings.ToLower(string(fields[1]))
 			continue
 		}
 		if bytes.Contains(fields[0], []byte("Version:")) {
@@ -181,14 +210,15 @@ func DebPkgInfo(ctx context.Context, path string) (*PkgInfo, error) {
 // InstallAptPackages installs apt packages.
 func InstallAptPackages(ctx context.Context, pkgs []string) error {
 	args := append(aptGetInstallArgs, pkgs...)
-	install := exec.Command(aptGet, args...)
-	install.Env = append(os.Environ(),
-		"DEBIAN_FRONTEND=noninteractive",
-	)
-	stdout, stderr, err := runner.Run(ctx, install)
+	cmdModifiers := []cmdModifier{
+		func(cmd *exec.Cmd) {
+			cmd.Env = append(os.Environ(), "DEBIAN_FRONTEND=noninteractive")
+		},
+	}
+	stdout, stderr, err := runAptGetWithDowngradeRetrial(ctx, args, cmdModifiers)
 	if err != nil {
 		if dpkgRepair(ctx, stderr) {
-			stdout, stderr, err = runner.Run(ctx, install)
+			stdout, stderr, err = runAptGetWithDowngradeRetrial(ctx, args, cmdModifiers)
 		}
 	}
 	if err != nil {
@@ -200,14 +230,15 @@ func InstallAptPackages(ctx context.Context, pkgs []string) error {
 // RemoveAptPackages removes apt packages.
 func RemoveAptPackages(ctx context.Context, pkgs []string) error {
 	args := append(aptGetRemoveArgs, pkgs...)
-	remove := exec.Command(aptGet, args...)
-	remove.Env = append(os.Environ(),
-		"DEBIAN_FRONTEND=noninteractive",
-	)
-	stdout, stderr, err := runner.Run(ctx, remove)
+	cmdModifiers := []cmdModifier{
+		func(cmd *exec.Cmd) {
+			cmd.Env = append(os.Environ(), "DEBIAN_FRONTEND=noninteractive")
+		},
+	}
+	stdout, stderr, err := runAptGet(ctx, args, cmdModifiers)
 	if err != nil {
 		if dpkgRepair(ctx, stderr) {
-			stdout, stderr, err = runner.Run(ctx, remove)
+			stdout, stderr, err = runAptGet(ctx, args, cmdModifiers)
 		}
 	}
 	if err != nil {
@@ -288,7 +319,7 @@ func AptUpdates(ctx context.Context, opts ...AptGetUpgradeOption) ([]*PkgInfo, e
 		return nil, err
 	}
 
-	out, err := run(ctx, aptGet, args)
+	out, _, err := runAptGetWithDowngradeRetrial(ctx, args, []cmdModifier{})
 	if err != nil {
 		return nil, err
 	}
@@ -298,7 +329,8 @@ func AptUpdates(ctx context.Context, opts ...AptGetUpgradeOption) ([]*PkgInfo, e
 
 // AptUpdate runs apt-get update.
 func AptUpdate(ctx context.Context) ([]byte, error) {
-	return run(ctx, aptGet, aptGetUpdateArgs)
+	stdout, _, err := runAptGet(ctx, aptGetUpdateArgs, []cmdModifier{})
+	return stdout, err
 }
 
 func parseInstalledDebpackages(data []byte) []*PkgInfo {
