@@ -14,8 +14,9 @@
 
 // Only build for linux but not on unsupported architectures.
 
+//go:build linux && (386 || amd64 || arm64)
 // +build linux
-// +build 386 amd64
+// +build 386 amd64 arm64
 
 package packages
 
