@@ -84,7 +84,7 @@ func runWithPty(cmd *exec.Cmd) ([]byte, []byte, error) {
 	cmd.SysProcAttr = &syscall.SysProcAttr{
 		Setctty: true,
 		Setsid:  true,
-		Ctty:    int(tty.Fd()),
+		Ctty:    0, // Stdin of the child process
 	}
 
 	var stdout bytes.Buffer
@@ -113,15 +113,26 @@ func runWithPty(cmd *exec.Cmd) ([]byte, []byte, error) {
 		}
 	}()
 
-	err = cmd.Run()
+	cmdErr := cmd.Run()
+
 	if err := tty.Close(); err != nil {
-		return nil, nil, err
+		return stdout.Bytes(), stderr.Bytes(), err
+
+	}
+	wg.Wait()
+
+	if cmdErr != nil {
+		// Yum returns non-zero exit values on non-error conditions.
+		// Errors which are *not* in this category indicate failure.
+		if _, ok := cmdErr.(*exec.ExitError); !ok {
+			return stdout.Bytes(), stderr.Bytes(), cmdErr
+
+		}
 	}
 
-	wg.Wait()
 	// Exit code 0 means no updates, 1 probably means there are but we just didn't install them.
-	if err == nil {
-		return nil, nil, err
+	if cmdErr == nil {
+		return nil, nil, nil
 	}
 	return stdout.Bytes(), stderr.Bytes(), retErr
 }
